[BACKLOG-31191] Revert "[BACKLOG-31191] Adjust github release creation to have the proper tag name" and move the TAG_NAME evaluation to when build data is processed for ease of use.
